refactor: move voting to the service
