+++ frontend/svelte/src/lib/services/proposals.services.ts
@@ -11,6 +11,7 @@ import {
 } from "@dfinity/nns";
 import { get } from "svelte/store";
 import { LIST_PAGINATION_LIMIT } from "../constants/constants";
+import { busyStore } from "../stores/busy.store";
 import { i18n } from "../stores/i18n";
 import {
   proposalsFiltersStore,
@@ -20,6 +21,8 @@ import {
 import { toastsStore } from "../stores/toasts.store";
 import { createAgent } from "../utils/agent.utils";
 import { enumsExclude } from "../utils/enum.utils";
+import { stringifyJson, uniqueObjects } from "../utils/utils";
+import { listNeurons } from "./neurons.services";
 
 export const listProposals = async ({
   clearBeforeQuery = false,
@@ -189,9 +192,8 @@ export const getProposalId = (path: string): ProposalId | undefined => {
 
 /**
  * Makes multiple registerVote calls (1 per neuronId).
- * @returns List of errors (order is preserved)
  */
-export const castVote = async ({
+export const registerVotes = async ({
   neuronIds,
   proposalId,
   vote,
@@ -201,16 +203,50 @@ export const castVote = async ({
   proposalId: ProposalId;
   vote: Vote;
   identity: Identity | null | undefined;
-}): Promise<Array<GovernanceError | undefined>> => {
+}): Promise<void> => {
   if (!identity) {
     throw new Error(get(i18n).error.missing_identity);
   }
 
+  busyStore.start("vote");
+
+  try {
+    await requestRegisterVotes({
+      neuronIds,
+      proposalId,
+      vote,
+      identity,
+    });
+  } catch (error) {
+    console.error("vote unknown:", error);
+    toastsStore.show({
+      labelKey: "error.register_vote_unknown",
+      level: "error",
+      detail: stringifyJson(error, { indentation: 2 }),
+    });
+  }
+
+  await listNeurons();
+  busyStore.stop("vote");
+};
+
+const requestRegisterVotes = async ({
+  neuronIds,
+  proposalId,
+  vote,
+  identity,
+}: {
+  neuronIds: bigint[];
+  proposalId: ProposalId;
+  vote: Vote;
+  identity: Identity;
+}): Promise<void> => {
   const governance: GovernanceCanister = GovernanceCanister.create({
     agent: await createAgent({ identity, host: process.env.HOST }),
   });
+
   // TODO: switch to Promise.allSettled -- https://dfinity.atlassian.net/browse/L2-369
-  return Promise.all(
+  const errors: Array<GovernanceError | undefined> = await Promise.all(
     neuronIds.map((neuronId) =>
       (
         governance.registerVote({
@@ -221,4 +257,18 @@ export const castVote = async ({
       ).then((res) => ("Err" in res ? res.Err : undefined))
     )
   );
+
+  // show only uqique error messages
+  const errorDetails = uniqueObjects(errors.filter(Boolean))
+    .map((error) => stringifyJson(error?.errorMessage, { indentation: 2 }))
+    .join("\n");
+
+  if (errorDetails.length > 0) {
+    console.error("vote:", errorDetails);
+    toastsStore.show({
+      labelKey: "error.register_vote",
+      level: "error",
+      detail: `\n${errorDetails}`,
+    });
+  }
 };
